# refine the gpu config for performance optimization