@wangchaochaohu 2020-11-02
> 这个修改除了会对这2个OP有提升，对其他的也会有吗？

有的